requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pytest
httpx
python-telegram-bot
reportlab